requirements.txt
Django==1.10rc1
django-filter==0.13.0
djangorestframework==3.4.0
Markdown==2.6.6
